+++ src/Installers/NodeDependencyInstaller.php
@@ -8,11 +8,11 @@ use Symfony\Component\Process\Process;
 
 class NodeDependencyInstaller extends AbstractInstaller
 {
-    public function handle(array $settings = []): void
+    public function handle(): void
     {
         $this->start('Installing node dependencies');
 
-        $dependencies = $settings['node'] ?? [];
+        $dependencies = $this->settings['node'] ?? [];
 
         $process = $this->getProcessRunner($dependencies);