+++ src/Installers/NodeScriptInstaller.php
@@ -13,7 +13,7 @@ class NodeScriptInstaller extends AbstractInstaller
         ];
     }
 
-    public function handle(array $settings = []): void
+    public function handle(): void
     {
         $this->start('Installing scripts');