+++ src/Installers/TailwindInstaller.php
@@ -17,7 +17,7 @@ class TailwindInstaller extends AbstractInstaller
         ];
     }
 
-    public function handle(array $settings = []): void
+    public function handle(): void
     {
         $this->start('Installing Tailwind');