+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] 2022-01-21
+### Fixed
+- Fixed a bug in `NodeDependencyInstaller` that was throwing an error when running `yarn`.
+- Fixed a bug in `InertiaInstaller` that wasn't placing the middleware in `Kernel` correctly.
+
+### Changed
+- The installer classes array has been moved into the `vitamin` config file.
+- Lots of refactoring to help with new tests.
+
 ## [0.2.0] 2022-01-09
 ### Changed
 - Tailwind 3 configuration.